respeciate.0.2

This is respeciate but with object classes...

sp_find_profile finds the codes of speciate profiles in the local speciate archive using supplied search terms.

sp_profile extracts a speciate profile from the local speciate archive.

When supplied a respeciate object, print manages its appearance.

When supplied a respeciate object, plot provides a basic plot output. This uses base function barplot; also see note.

When supplied a respeciate.ref object, print manages its appearance.

sp_find_profile(term, by = "Keywords")

sp_profile(code)

# S3 method for respeciate
print(x, n = 6, ...)

# S3 method for respeciate
plot(x, n = NULL, order = TRUE, ..., legend.text = NULL, args.legend = NULL)

# S3 method for respeciate.ref
print(x, ...)

Arguments

term
character, the search term to use when searching archive.

by
character, the section of the archive to search, by default 'keywords'.

code
character or numeric, the speciate code of the required source profile (EPA SPECIATE term PROFILE_CODE).

x
the respeciate or respeciate.ref object to be printed, plotted, etc.

n
when plotting or printing a multi-profile, the number(s) of profile(s) to report.

...
any extra arguments, mostly ignored except by plot which passes them to barplot.

order
logical, order the species in the profile(s) by molecular weight before plotting.

legend.text, args.legend
barplot arguments.

Value

sp_find_profile returns a object of respeciate.ref class, a data.frame of profile information.

sp_profile returns a object of respeciate class, a data.frame containing a speciate profile.

Note

respeciate objects revert to data.frames when not doing anything package-specific, so you can still use as previously with lattice or ggplot2, useful it you are pulling multiple profiles and you exceed the base barplot capacity...

References

Simon, H., Beck, L., Bhave, P.V., Divita, F., Hsu, Y., Luecken, D., Mobley, J.D., Pouliot, G.A., Reff, A., Sarwar, G. and Strum, M., 2010. The development and uses of EPA SPECIATE database. Atmospheric Pollution Research, 1(4), pp.196-206.

Examples

if (FALSE) {
profile <- "Ethanol"
dt <- sp_find_profile(profile)
}
if (FALSE) {
x <- sp_profile(c(8833, 8850))
plot(x)
}