respeciate.generics

Generic functions for use with respeciate object classes.

as.respeciate(x, ...)

# Default S3 method
as.respeciate(x, ...)

# S3 method for class 'respeciate'
print(x, n = 6, ...)

# S3 method for class 'rsp_pls'
print(x, n = NULL, ...)

# S3 method for class 'respeciate'
plot(x, ...)

# S3 method for class 'rsp_pls'
plot(x, ...)

# S3 method for class 'respeciate'
summary(object, ...)

# S3 method for class 'respeciate'
merge(x, y, ...)

Arguments

x

the respeciate object to be printed, plotted, etc.

...

any extra arguments, mostly ignored except by plot which passes them to rsp_plot_profile and merge with passes them to merge.

n

when plotting or printing a multi-profile object, the maximum number of profiles to report.

object

like x but for summary.

y

a second data set, typically a data.frame or a respeciate object, to be merged with x

Value

These generic functions/methods generate typical outputs for respeciate data sets and models: When supplied a data.frame or similar, as.respeciate attempts to coerce it into a respeciate object.

When supplied a respeciate object, print manages its appearance. When supplied a respeciate object, plot provides a basic plot output. This is currently wrapper for the respeciate function rsp_plot_profile.

When supplied a respeciate object, summary generates a summary table of profile information.

When supplied a respeciate object and a second respeciate-like object, e.g. data.frame, respeciate object, etc, merge attempts to merge them using common data columns. You can refine the merge operation using additional arguments.

Note

respeciate objects revert to data.frames when not doing anything package-specific, so you can still use them like data.frames with other packages. This is useful if you have other ideas how to plot more complex (multiple-profile, multiple-species) data sets, and want to use graphics packages like lattice or ggplot2.