Getting archived profiles

Getting source profile(s) from the local respeciate archives.

rsp(..., include.refs = FALSE, source = "all")

rsp_profile(...)

Arguments

...

The function assumes all inputs (except include.refs and source) are profile identifiers: namely, PROFILE_CODE and Species.Id in SPECIATE and SPECIEUROPE, respectively, or potential sources of profile information and requests these form the local respeciate archives. Typically, simple objects like character and numeric vectors, as assumed to be profile identifiers and composite data-types like respeciate or data.frame objects are assumed to contain a column named .profile.id, the respeciate equivalent of PROFILE_CODE and Species.Id. All recovered identifiers are requested and unrecognized ids (and duplicates) are ignored.

include.refs

logical, if profile reference information should be included when extracting the requested profile(s) from the archive, default FALSE.

source

character, the local archive to request a profile from: 'us' US EPA SPECIATE, 'eu' EU JRC SPECIEUROPE, or 'all' (the default) both.

Value

rsp_profile or the short-hand rsp return an object of respeciate class, a data.frame containing one or more profile from the local respeciate archive.

Note

The option include.refs adds profile source reference information to the returned respeciate data set. The default option is to not include these because some SPECIATE profiles have several associated references and including these replicates records, once per reference. respeciate code is written to handle this but if you are developing own methods or code and include references in any profile build you may be biasing some analyses in favor of those multiple-reference profile unless you check and account such cases.

References

For SPECIATE:

Simon, H., Beck, L., Bhave, P.V., Divita, F., Hsu, Y., Luecken, D., Mobley, J.D., Pouliot, G.A., Reff, A., Sarwar, G. and Strum, M., 2010. The development and uses of EPA SPECIATE database. Atmospheric Pollution Research, 1(4), pp.196-206.

For SPECIEUROPE:

Pernigotti, D., Belis, C.A., Spano, L., 2016. SPECIEUROPE: The European data base for PM source profiles. Atmospheric Pollution Research, 7(2), pp.307-314. DOI: https://doi.org/10.1016/j.apr.2015.10.007

See also

Examples

if (FALSE) {
x <- rsp_profile(8833, 8850)
plot(x)}