Data averaging multiple profile data sets

Functions to build composite respeciate profiles

rsp_average_profile generates an average composite of a supplied multi-profile respeciate object.

rsp_average_profile(rsp, code = NULL, name = NULL, method = 1, ...)

Arguments

rsp

A respeciate object, a data.frame of respeciate profiles.

code

required character, the unique profile code to assign to the average profile.

name

character, the profile name to assign to the average profile. If not supplied, this defaults to a collapsed list of the codes of all the profiles averaged.

method

numeric, the averaging method to apply: Currently only 1 (default) mean(rsp).

...

additional arguments, currently ignored

Value

rsp_average_profile returns a single profile average version of the supplied respeciate profile.

Note

In development function; arguments and outputs likely to be subject to change.

This is one of the very few respeciate functions that modifies the WEIGHT_PERCENT column of the respectiate data.frame.