plotting respeciate source profiles

General plots for respeciate objects.

rsp_plot functions generate plots for supplied respeciate data sets.

rsp_plot_profile(
  rsp,
  id,
  multi.profile = "group",
  order = TRUE,
  log = FALSE,
  ...,
  silent = FALSE,
  output = "default"
)

rsp_plot_species(
  rsp,
  id,
  multi.species = "group",
  order = FALSE,
  log = FALSE,
  ...,
  silent = FALSE,
  output = "default"
)

rsp_plot_match(
  rsp,
  ref = NULL,
  plot.type = 2,
  log = FALSE,
  ...,
  output = "plot"
)

Arguments

rsp

A respeciate object, a data.frame of respeciate profiles.

id

numeric, the indices of profiles or species to use when plotting with rsp_plot_profile or rsp_plot_species, respectively. For example, rsp_plot_profile(rsp, id=1:6) plots first 6 profiles in respeciate object rsp.

multi.profile

character, how rsp_plot_profile should handle multiple profiles, e.g. 'group' or 'panel' (default group).

order

logical, order the species in the profile(s) by relative abundance before plotting.

log

logical, log y scale when plotting.

...

any additional arguments, typically passed on the lattice plotting functions.

silent

logical, hide warnings when generating plots (default FALSE)

output

character, output method, one of: 'plot' to return just the requested plot; 'data' to return just the data; and, c('plot', 'data') to plot then return the data invisibly (default).

multi.species,

character, like multi.profile in rsp_plot_profile but for species in rsp_plot_species.

ref

respeciate or similar data set of profiles, used by rsp_match_plot as a reference when comparing with rsp. See rsp_match_profile for further details and other matching arguments.

plot.type

numeric, option if the rsp_plot... function includes different plot reports.

Value

rsp_plot graph, plot, etc usually as a trellis object.

Note

These functions are currently in development, so may change.

References

Most respeciate plots make extensive use of lattice and latticeExtra code:

Sarkar D (2008). Lattice: Multivariate Data Visualization with R. Springer, New York. ISBN 978-0-387-75968-5, http://lmdvr.r-forge.r-project.org.

Sarkar D, Andrews F (2022). latticeExtra: Extra Graphical Utilities Based on Lattice. R package version 0.6-30, https://CRAN.R-project.org/package=latticeExtra.

They also incorporate ideas from loa:

Ropkins K (2023). loa: various plots, options and add-ins for use with lattice. R package version 0.2.48.3, https://CRAN.R-project.org/package=loa.