Plots for use with (re)SPECIATE profile Positive Least Squares models

The pls_plot functions are intended for use with PLS models built using rsp_pls_profile (documented separately). They generate some plots commonly used with source apportionment model outputs.

pls_plot(pls, plot.type = 1, ...)

pls_plot_profile(pls, plot.type = 1, log = FALSE, ...)

pls_plot_species(pls, id, plot.type = 1, ...)

Arguments

pls

A sp_pls_profile output, intended for use with pls_ functions.

plot.type

numeric, the plot type if multiple options are available.

...

other arguments, typically passed on to the associated lattice plot.

log

(for pls_plot_profile only) logical, if TRUE this applies 'log' scaling to the primary Y axes of the plot.

id

numeric or character identifying the species or profile to plot. If numeric, these are treated as indices of the species or profile, respectively, in the PLS model; if character, species is treated as the name of species and profile is treated as the profile code. Both can be concatenated to produce multiple plots and the special case id = -1 is a short cut to all species or profiles, respectively.

Value

pls_plots produce various plots commonly used in source apportionment studies.