sp_ functions

sp_ functions for use with (re)speciate data in R...

sp_find functions search main data sets in the (re)SPECIATE archive using supplied search terms. sp_find_profile searches for profile records and sp_find_species searches for species records.

sp_profile extracts a SPECIATE profile from the local (re)SPECIATE archive.

sp_find_profile(..., by = "keywords", partial = TRUE)

sp_find_species(..., by = "species_name", partial = TRUE)

sp_profile(code)

Arguments

...

for sp_find functions, character(s), any search term(s) to use when searching the local (re)SPECIATE archive for relevant records.

by

character, the section of the archive to search, by default 'keywords' for sp_find_profile and 'species_names' for sp_find_species.

partial

logical, if TRUE (default) sp_find functions use partial matching.

code

character or numeric, the SPECIATE code of the required profile (EPA SPECIATE term PROFILE_CODE).

Value

sp_find_profile returns a object of respeciate.ref class, a data.frame of profile information.

sp_find_species returns a object of respeciate.spcs class, a data.frame of species information.

sp_profile returns a object of respeciate class, a data.frame containing a speciate profile.

References

Simon, H., Beck, L., Bhave, P.V., Divita, F., Hsu, Y., Luecken, D., Mobley, J.D., Pouliot, G.A., Reff, A., Sarwar, G. and Strum, M., 2010. The development and uses of EPA SPECIATE database. Atmospheric Pollution Research, 1(4), pp.196-206.

Examples

if (FALSE) {
profile <- "Ethanol"
dt <- sp_find_profile(profile)
dt}

if (FALSE) {
species <- "Ethanol"
sp <- sp_find_species(species)
sp}

if (FALSE) {
x <- sp_profile(c(8833, 8850))
plot(x)}