Find nearest matches from reference set of profiles

sp_match_profile compares a supplied species (re)SPECIATE profile and a reference set of supplied profiles and attempt to identify nearest matches on the basis of correlation coefficient.

sp_match_profile(x, ref, matches = 10, rescale = 5, min.n = 8, test.x = FALSE)

Arguments

x

A respeciate object or similar data.frame containing a species profile to be compared with profiles in ref. If x contains more than one profile, these are averaged (using sp_average_profile), and the average compared.

ref

A respeciate object, a data.frame containing a multiple species profiles, to be used as reference library when identifying nearest matches for x.

matches

Numeric (default 10), the maximum number of profile matches to report.

rescale

Numeric (default 5), the data scaling method to apply before comparing x and profiles in ref: options 0 to 5 handled by sp_rescale.

min.n

numeric (default 8), the minimum number of paired species measurements in two profiles required for a match to be assessed. See also sp_species_cor.

test.x

Logical (default FALSE). The match process self-tests by adding x to ref, which should generate a perfect fit=1 score. Setting test.x to TRUE retains this as an extra record.

Value

sp_match_profile returns a fit report: a data.frame of up to n fit reports for the nearest matches to x from the reference profile data set, ref.